[SW-91][refactor]레이아웃 수정 전 준비
css 수정 전 계층관계를 명확히 하기 위해
Options, Panel, ScreenShare 별도 파일로 분리했습니다.
