+++ front/src/spaceMain/Navigation.tsx
@@ -22,6 +22,9 @@ import {useContext} from 'react';
 import GlobalContext from './GlobalContext';
 import MicOnOff from './MicOnOff';
 import Profile from './Profile';
+import ScreenShare from './ScreenShare';
+import Options from './Options';
+import Panel from './Panel';
 
 interface NavigationProps {
   initialInfo: [avatarIdx: number, nickname: string];
@@ -34,43 +37,6 @@ interface NavigationProps {
 const Navigation = (props: NavigationProps) => {
   const globalContext = useContext(GlobalContext);
 
-  const panel = () => {
-    const onClickLink = () => {
-      // 복사 시켜줘야한다. 성공했는지 못했는지는 분기해주자.
-      message.success('복사 성공!');
-    };
-    const onClickUsers = () => {
-      console.log('user!!');
-    };
-    const onClickMessage = () => {
-      console.log('message!!!');
-    };
-    return (
-      <Menu>
-        <Menu.Item key="0">
-          <button onClick={onClickLink}>참여 링크 복사</button>
-        </Menu.Item>
-        <Menu.Item key="1">
-          <button onClick={onClickUsers}>사용자</button>
-        </Menu.Item>
-        <Menu.Item key="3">
-          <button onClick={onClickMessage}>메시지</button>
-        </Menu.Item>
-      </Menu>
-    );
-  };
-
-  const screenshare = (
-    <Menu>
-      <Menu.Item key="0">
-        <a href="https://www.antgroup.com"> 전체 화면</a>
-      </Menu.Item>
-      <Menu.Item key="1">
-        <a href="https://www.aliyun.com"> 창</a>
-      </Menu.Item>
-    </Menu>
-  );
-
   const exit = () => {
     if (globalContext.peerManager) {
       globalContext.peerManager.close();
@@ -78,44 +44,36 @@ const Navigation = (props: NavigationProps) => {
     props.goToHome();
   };
 
-  const options = (
-    <Menu>
-      <Menu.Item key="0">
-        <a href="https://www.antgroup.com">설정</a>
-      </Menu.Item>
-      <Menu.Item key="1">
-        <a href="https://www.aliyun.com"> 문제 해결 및 도움말</a>
-      </Menu.Item>
-    </Menu>
-  );
-
   return (
-    <div id="footer">
-      <Profile />
-      <span className="footerCenter">
-        <div style={{display: 'inline'}}></div>
-        <MicOnOff />
-        <Dropdown overlay={screenshare} trigger={['click']}>
-          <a className="ant-dropdown-link" onClick={e => e.preventDefault()}>
-            <DesktopOutlined className="navigationObject" />
-          </a>
-        </Dropdown>
-        <Dropdown overlay={options} trigger={['click']}>
-          <a className="ant-dropdown-link" onClick={e => e.preventDefault()}>
-            <EllipsisOutlined
-              className="navigationObject"
-              style={{transform: 'rotate(90deg)'}}
-            />
-          </a>
-        </Dropdown>
-        <LogoutOutlined className="navigationObject" onClick={exit} />
-      </span>
-      <Dropdown overlay={panel} trigger={['click']}>
-        <a className="ant-dropdown-link" onClick={e => e.preventDefault()}>
-          <UpOutlined className="navigationObject" />
-        </a>
-      </Dropdown>
-    </div>
+    <nav className="navbar">
+      <div className="navbar_profile">
+        <Profile />
+      </div>
+
+      <ul className="navbar_menu">
+        <li>
+          <MicOnOff />
+        </li>
+
+        <li>
+          <ScreenShare />
+        </li>
+
+        <li>
+          <Options />
+        </li>
+
+        <li>
+          <LogoutOutlined className="navigationObject" onClick={exit} />
+        </li>
+      </ul>
+
+      <ul className="navbar_panel">
+        <li>
+          <Panel />
+        </li>
+      </ul>
+    </nav>
   );
 };
