+++ front/src/spaceMain/ScreenShare.tsx
@@ -1,0 +1,27 @@
+import React from 'react';
+import {Menu, Dropdown} from 'antd';
+import {DesktopOutlined} from '@ant-design/icons';
+
+const ScreenShare = () => {
+  const screenshare = () => {
+    return (
+      <Menu>
+        <Menu.Item key="0">
+          <a href="https://www.antgroup.com"> 전체 화면</a>
+        </Menu.Item>
+        <Menu.Item key="1">
+          <a href="https://www.aliyun.com"> 창</a>
+        </Menu.Item>
+      </Menu>
+    );
+  };
+  return (
+    <Dropdown overlay={screenshare} trigger={['click']}>
+      <a className="ant-dropdown-link" onClick={e => e.preventDefault()}>
+        <DesktopOutlined className="navigationObject" />
+      </a>
+    </Dropdown>
+  );
+};
+
+export default ScreenShare;
