+++ front/src/spaceMain/spaceMain.css
@@ -1,69 +1,30 @@
-#footer {
-  position: absolute;
-  left: 0;
-  bottom: 0;
-  height: 10vh;
-  width: 100%;
-  padding: 0px;
-  text-align: center;
-  color: white;
-  background: #bae7ff;
+/* navbar 전체 */
+.navbar {
   display: flex;
   justify-content: space-between;
-  font-size: 6vh;
-  z-index: 11;
+  align-items: center;
+  background-color: white;
+  padding: 8px 12px;
 }
 
-.navigationObject {
-  padding: 0px 10px 0px 10px;
-  color: #00a0ff;
-  text-align: center;
-}
+/* navbar 왼쪽 부분 */
 
-#spaceMain {
-  margin: 0%, 0%, 0%, 0%;
-  padding-left: 0%;
-  padding-top: 0%;
-  padding-right: 0%;
-  height: 90vh;
-}
-
-button {
-  background: none;
-  color: inherit;
-  border: none;
-  padding: 0;
-  font: inherit;
-  cursor: pointer;
-  outline: inherit;
-}
-
-.profile {
-  padding: 10px 10px 10px 10px;
-}
-
-.profile > div {
-  padding: 5px 0px 5px 0px;
-}
-
-.avatar:not(div) {
+/* navbar 중간부분 */
+.navbar_menu {
   display: flex;
-  justify-content: space-between;
+  list-style: none;
+  padding-left: 0;
 }
-/* style={{color: '#00a0ff', paddingLeft: '10px'}} */
-
-.avatar button {
-  color: #00a0ff;
-  padding-left: 6px;
-  padding-right: 6px;
+.navbar_menu li {
+  padding: 8px 12px;
 }
 
-.profileDisplay {
-  font-size: 15px;
-  font-weight: bold;
+/* navbar 오른쪽 부분 */
+.navbar_panel {
+  list-style: none;
 }
-/* style={{color: '#00a0ff', paddingLeft: '10px'}} */
 
+/* joystick */
 .joystickBase {
   position: absolute;
   visibility: hidden;
